Initial Proof of Concept
Added basic chrome extension set up. We use webpack to build our JS files.

I was getting a CORS error trying to access the Maps API directly in the content layer. This was due to a chrome change in policy preventing external API calls in the content layer.  Instead we make our API calls in the background layer and send messages across.

Google Maps/Place API was not able to give me a restaurant name from only an address, despite the API object looking like it would. Instead it returned the address as the name. As a workaround I scrape the latitude and longitude and look for restaurants within 50 meters of those values. We then use some logic to filter for exact addresses and then do a string comparison to identify matches.

I added some simple styles to explain to the user why the restaurant was or was not a ghost kitchen.