Fixed matching logic to handle address and restaurant name matching more accuratelyt
Made the system for matching addresses and restaurants more specific. Because there was no convenient library to attempt to standardize the text available we were forced to create a hackier system. We attempt to iteratively apply the changes so that the majority of cases are handled before the workarounds have to be applied. The choices here opens us up to specific bugs but it should perform better in almost every ordinary case. And we will try to fail showing the user the restaurant or the candidate restaurants along with the google address link so that they can spot confirm any cases that slip through.

Added regex files

Cleaned console.logs and unused functions. Renamed some functions. Removed unneeded permissions. Added comments liberally.

And Added a potentially haunted state message and style