Fixed issue where script would only only first page load
Content_scripts should load and run whenever we hit any of the pages listed in the "matches" in the manifest.json. However, I was finding that if I loaded a page that was in the matches and then navigated to another page in the matches, it would not rerun the script. I added a listener that would see if the user changed to one of the restaurant pages, so that we could run the check another time.

When doing this I found that the script could be called over and over again, so I added a mutex to prevent the script from running more than once at a time. I also added a check to bail if we had already added a badge to the DOM.

I expanded the range to 50 meters to address specific restaurant bug cases I happened across.

Remaned helpers.ts to HTMLelements.ts

Removed unused popup HTML.